requirements.txt
python-Levenshtein
numpy~=1.26.2
nltk~=3.8.1
Levenshtein~=0.23.0
torch~=2.1.2
pytorch_lightning~=2.1.3
Pillow~=10.1.0
transformers~=4.36.2
sentencepiece~=0.1.99
protobuf~=4.25.1
